refactor: Cleanup duplicate types and split large files (#159)
Phase 1: Type Consolidation
- Remove duplicate FinalResultData from tracking.rs and state.rs
- Consolidate to canonical FinalResult from update.rs
- Update all consumers (CLI, metrics) to use nested metadata paths

Phase 2: Remove Deprecated Aliases
- Remove SubQueryTiming, SubQueryState, SubQueryStatus from gemicro-runner

Phase 3: File Splits
- Split llm.rs (1133 lines) into llm/{mod,client,request}.rs
- Split trajectory.rs (927 lines) into trajectory/{mod,data,builder}.rs

Additional cleanup from review:
- Add FinalResult/ResultMetadata to CLAUDE.md canonical import table
- Remove rust-genai re-exports (violated Single Source of Truth principle)
- Remove redundant from_state() wrapper method
- Remove unreachable wildcard pattern in metrics.rs

Closes #159
